allow float sized bullet

+++ pyrl/tasks/pyale/games/london/object.py
@@ -388,10 +388,10 @@ class Bullet(pygame.sprite.Sprite):
 
 		self.image = Bullet.image
                 self.baseImageSize = self.image.get_size()
-                bullet_size = int(get_val(os.environ, 'BULLET', 1.))
+                bullet_size = float(get_val(os.environ, 'BULLET', 1.))
                 self.image = pygame.transform.scale(self.image,
-                        (self.baseImageSize[0] * bullet_size, 
-                         self.baseImageSize[1] * bullet_size))
+                        (int(self.baseImageSize[0] * bullet_size), 
+                         int(self.baseImageSize[1] * bullet_size)))
 		self.rect = self.image.get_rect(center=pos)
 		self.coords = [float(self.rect.centerx), float(self.rect.centery)]
 		self.angle = self.get_angle(pos, aimAt)
